app/requirements.txt
flask
pdfplumber
python-docx
openai
python-dotenv
groq

